test:(test_gwf_utl) : renamed two of the autotests
Renamed the binary file and time series autotests to be of the form test_gwf_utlxx_purpose.  These tests are testing the utility parts of the code, so they are grouped now into a gwf_utl category.  We should continue to name our tests in a consistent manner so that we use wildcards with nosetests to test only the things we are working on.