feat(GwtSrbModule) : finished up much of the Sorption Package structure
Sorption package is up and running and has a test, input instructions, and flopy support.  Right now it only has equilibrium-control linear sorption.  Should also add Freundlich, Langmuir and the nonequilibrium options that are in MT3D.  And what about dual-domain?  Does that go here?  And there may still be a question about the volume that should be used for calculating sorption.  Right now, we are using Vcell, but this might need another look.  Also still need to add and test observations.