fix(GwfModule): fix to apply ptc to problems with the STO package
Fix to apply ptc during steady-state stress periods to models using
the Newton-Raphson formulation and including the STO package.
Previously ptc was only applied to models using the Newton-Raphson
formulation and no STO package - this is not the intended behaviour.
Also migrated DEV_NO_PTC option to NO_PTC option. Added autotest
(test_gwf_ptc01) that tests that the same results are achieved for
problems with and without the STO packages (this test is identical to
MODFLOW-NWT Problem 3 - low recharge).

closes #47