Transport (#63)
* feat(GwtSrbModule): created new sorption package for transport model

First crack at the sorption package.  The GwtSrbModule has been implemented, but has not yet been added to the gwt model.  The source compiles and existing tests run and pass so the new code doesn't break anything.

* feat(GwtSrbModule) : finished up much of the Sorption Package structure

Sorption package is up and running and has a test, input instructions, and flopy support.  Right now it only has equilibrium-control linear sorption.  Should also add Freundlich, Langmuir and the nonequilibrium options that are in MT3D.  And what about dual-domain?  Does that go here?  And there may still be a question about the volume that should be used for calculating sorption.  Right now, we are using Vcell, but this might need another look.  Also still need to add and test observations.

* test(GwtSrbModule): removed a hard coded assert False

Removed a statement that was left in by accident.

* docs(readme): fix directory name

* fix(GwfModule): fix to apply ptc to problems with the STO package

Fix to apply ptc during steady-state stress periods to models using
the Newton-Raphson formulation and including the STO package.
Previously ptc was only applied to models using the Newton-Raphson
formulation and no STO package - this is not the intended behaviour.
Also migrated DEV_NO_PTC option to NO_PTC option. Added autotest
(test_gwf_ptc01) that tests that the same results are achieved for
problems with and without the STO packages (this test is identical to
MODFLOW-NWT Problem 3 - low recharge).

closes #47

* refactor(pre-commit.py): use OrderdedDict to load and write code.json

Updated the code.json writer to use an ordered dictionary so that the order is always the same.  This will minimize the number of changes that appear in the changelog due to ordering changes that do not change content.

* refactor(InputOutputModule): refactor dclosetest (#49)

Refactor dclosetest and replace with IS_SAME in GenericUtilities.
Convert IS_SAME to logical function and allow passing of evaluation
value (eps). If eps is not passed then DSAME is used. Modified calls
to dclosetest to IS_SAME but did not pass an eps value (unlike what
was done previously). All tests pass so current tests are not sensitive
to a passed eps value. Need to monitor this for timeseries functionality
which used dclosetest.

* refactor(pre-commit.py): use OrderdedDict to load and write code.json (#50)

Updated the code.json writer to use an ordered dictionary so that the order is always the same.  This will minimize the number of changes that appear in the changelog due to ordering changes that do not change content.

* perf(GwfModule): improvements to ptc (#53)

Small improvements to ptc to reduce the ptcdel value loaded on the
diagonal.

* fix(ObsModule): implemented non-advancing output (#55)

Modified the pbservation process to use non-advancing output instead of
fixed length strings when writing ascii output. The previous use of
fixed length strings resulted in truncation of ascii observation output
when the product of user-specified digits + 7 and the number of
observations exceeded 5000.

Closes #54

* fix(GwfGwfExchangeModule): correct specific discharge calculation for LGR (#56)

* refactor(pre-commit.py): use OrderdedDict to load and write code.json

Updated the code.json writer to use an ordered dictionary so that the order is always the same.  This will minimize the number of changes that appear in the changelog due to ordering changes that do not change content.

* fix(GwfGwfExchangeModule): specific discharge not updated correctly for LGR

The sign for the face normal needed to be flipped for model 2 so that specific discharge would be calculated correctly.  Added a test to make sure that the velocities are correct.  Addresses #51.

* fix(GwfGwfExchangeModule): specific discharge not updated correctly for LGR

Updated the release notes and added some docstrings to the autotest.

* docs(dfn): fix 3 dfn typos (#59)

- `double` → `double precision`
- `valid_values` → `valid`
- `in_record = false` → `in_record false`

* docs(mf6ivar): readme.md updated to include valid keyword (#60)

* refactor(pre-commit.py): use OrderdedDict to load and write code.json

Updated the code.json writer to use an ordered dictionary so that the order is always the same.  This will minimize the number of changes that appear in the changelog due to ordering changes that do not change content.

* fix(GwfGwfExchangeModule): specific discharge not updated correctly for LGR

The sign for the face normal needed to be flipped for model 2 so that specific discharge would be calculated correctly.  Added a test to make sure that the velocities are correct.  Addresses #51.

* fix(GwfGwfExchangeModule): specific discharge not updated correctly for LGR

Updated the release notes and added some docstrings to the autotest.

* docs(mf6ivar): readme.md needed a description of the valid keyword

* fix(ssm): added trap for case where flows are not available to transport model