fix(inputoutput): columns and width were not be checked properly for the print_format (#72)
Fixes #69.

This patch also includes some changes to the OutputControlData object so that it can be used directly in a package other than the output control package.  This new functionality is being used for the transport model as a way to save immobile domain concentrations in a manner similar to how heads are saved.