refactor(mover): refactor and clean up water mover for possible BOUNDNAME support (#86)
This is some minor clean up to the water mover that improves some of the comments, and eliminates storage of a 20-character string for each water mover, and instead replaces it with an integer.  Reading of the reach id is now done using the IO utility, extract_idnum_or_bndname.