refactor(BndType): Generalized BndType pointer to NPF ICELLTYPE to be model agnostic (#93)
ICELLTYPE is a member of BndType, which is supposed to be a general package class for GWF and other models.  But ICELLTYPE is an NPF thing.  This PR generalizes the ICELLTYPE member by allowing packages to set the origin of this variable.  For the GWF Boundary packages, the ictorigin value is set to "NPF".  If ictorigin is not set, then BndType will not try to set ICELLTYPE, and it will remain null().

Closes #67