refactor(tests): merged in develop and updated scripts to work with new flopy syntax
* feat(GwtSrbModule): created new sorption package for transport model

First crack at the sorption package.  The GwtSrbModule has been implemented, but has not yet been added to the gwt model.  The source compiles and existing tests run and pass so the new code doesn't break anything.

* feat(GwtSrbModule) : finished up much of the Sorption Package structure

Sorption package is up and running and has a test, input instructions, and flopy support.  Right now it only has equilibrium-control linear sorption.  Should also add Freundlich, Langmuir and the nonequilibrium options that are in MT3D.  And what about dual-domain?  Does that go here?  And there may still be a question about the volume that should be used for calculating sorption.  Right now, we are using Vcell, but this might need another look.  Also still need to add and test observations.

* test(GwtSrbModule): removed a hard coded assert False

Removed a statement that was left in by accident.

* fix(ssm): added trap for case where flows are not available to transport model

* feat(gwt): first and zero order reactions implemented plus other minor changes

Also added additional checks to make sure that ssm is used if there are sources/sinks in the model; fixed a problem with pointers in the dispersion package (we need to revisit the memory manager for the mem_setptr case because the underlying mt object isn't being updated if a pointer is reset.

* Cleanup on reactions in preparation for getting dual domain into the code.

* fix(rct): fixed porosity error in sorbtion equation

* feat(srb): dual domain formulate is implemented but results are wrong and budget not implemented yet

* feat(srb): dual domain mass transfer working now including budget

Still need to break the budget into the individual components, not just the transfer to the immobile domain

* feat(imd): first commit of immobile domain package (IMD)

By making the IMD Package a BndType Package, users can add as many immobile domains as desired.  This means that triple porosity systems can be simulated by specifying two immobile domain packages.  Or the user can specify n number of IMD Packages.  Still more to do.  Need ability to output immobile domain concentration.  Need to refactor the Reactions Package to remove the immobile domain capability.  Need to update the definition files and add one for IMD.  Need to add tests.

* feat(imd): added printing of an immobile domain budget to the listing file

* feat(rct): changed SRB to RCT and added dfn for IMD Package

* feat(rct): added source file for reaction package

* adding to the todo list

* feat(dfn): added disu dfn for transport

* fix(ci): fix to pre-commit when working on a fork

* fix(update_flopy): there is a new flopy.dfn file that cannot be removed (#112)

* refactor(tests): to work with recent flopy changes (#113)

* refactor(tests): to work with recent flopy changes

* fix(test_gwf_utl02): added the time series filename due to possible bug in flopy

* fix(autotest): updated scripts to use new flopy syntax