eliminate gwt dependence on gwfbndlist and replaced with new PackageBudgetType
this is a big commit which pushes closer to having full capabilities to run gwt using budget and head files from a previous gwf run.  The PackageBudget objects are stored in an array in FMI, with one entry for each flow package.  The PackageBudget object can get its information from a gwf model running in the current simulation or it can get it from a budget file.