feat(fmi): working on running transport from saved gwf flows (#208)
* feat(fmi): working on running transport from saved gwf flows

* added a simple head file reader to fmi

* Added the FMI mf6io information

FMI now described in the mf6io documention.  The FMI definition file should now allow FMI to be created by flopy.

* Implemented head file reader to move closer toward all gwt input from binary files

* forgot to shut off the assert False

* modified the budget file reader to store a list of budget terms

* eliminate gwt dependence on gwfbndlist and replaced with new PackageBudgetType

this is a big commit which pushes closer to having full capabilities to run gwt using budget and head files from a previous gwf run.  The PackageBudget objects are stored in an array in FMI, with one entry for each flow package.  The PackageBudget object can get its information from a gwf model running in the current simulation or it can get it from a budget file.

* GWT can now be run from a GWF model in memory or with FMI loading head and budget from binary files

* Added the flow imbalance correction as an option

* updated a few tex files for fmi