feat(maxerrors): limit the number of errors stored to maxerrors (#221)
Error messages are stored in memory.  If something goes terribly wrong, the number of errors can be so large that memory runs out.  Limit the number of errors stored, by default, to 1000.  The user can control this maximum number by specifying MAXERRORS in the mfsim.nam options block.  Number of errors and number of errors above maxerrors are not written to mfsim.lst and to terminal.

Closes #106