feat(cmask): add mask for calculation of amat terms (#215)
* feat(cmask): skip calculation of amat terms in *_fc and *_fn for masked connections

The purpose of this PR is to allow another package or module to shutoff the flow connection between two cells.  A connection can be masked by calling set_mask with the position within the ja and amat arrays of the masked connection.  A value of zero means the connection is masked.  If not used, the masked array points to ja, which has non-zero values for all connections.  The first call to set_mask allocates new memory for the masked array of size nja and initializes all values to one.