fix(gwt): fix error when using gwt and disu when vertices not specified
Data for calculating normal vector and unit vector components are
not available with DISU grids if VERTEX data are not specified.
Also modified GWF-GWT exchange to reset integer flag to calculate
specific discharge data (icalcspdis=1) only if the GWT dispersion
package is on. Previously, the specific discharge integer flag was
reset to 1 anytime the GWF-GWT exchange was used. This resulted in
segfault issues for DISU models without VERTEX data but only used
advection (ADV).