feat(csub): add hbar to equations
Update zcell and stress calculations to use hbar. Refactor
csub_calc_znode function. Refactor coarse-grained compaction and add
autotest for non-newton and newton that includes a cell that dries and
resaturates. Refactor no-delay interbed compaction and add
autotest for non-newton and newton that includes a cell that dries and
resaturates.