feat(csub): add hbar to equations (#614)
Update zcell and stress calculations to use hbar. Refactor
csub_calc_znode function. Refactor coarse-grained compaction and add
autotest for non-newton and newton that includes a cell that dries and
resaturates. Refactor no-delay interbed compaction and add
autotest for non-newton and newton that includes a cell that dries and
resaturates. Removed saturation from delay bed exchange with aquifer
term. Added test for delay beds that includes a cell that dries and
resaturates.