fix(NumericalSolution): revert fix for issue 655
Previous changes to address issue 655 are not necessary as a result
of changes to linear accelerators to trap potential divide by zero
errors and exit the linear iteration loop. Partial refactor of
comments in ims8linear.f90 to be consistent with rest of MODFLOW 6 code.
Added some doxygen comments to ims8linear.f90.