docs(fmi.tex): options block to PACKAGEDATA block (#705)
I changed the word "options" to "PACKAGEDATA". It previously said "Flows are provided from a previous GWF model simulation—in this case FMI should be provided in
the GWT name file and the head and budget files should be listed in the FMI options block." But these files are actually provided in the PACKAGEDATA block. I used uppercase because that is consistent with its usage two bullet points up.