feat(ats): adaptive time stepping (#676)
A new Adaptive Time Step (ATS) utility was added.  The ATS utility allows any stress period to be overridden with an alternative time stepping approach.  The ATS utility implements two main capabilities (1) the capability to retry failed time steps with a shorter time step repeatedly until convergence is achieved, and (2) the capability to shorten and lengthen time steps based on simulation behavior.  These capabilities are described in the user input and output guide in a new section on the ATS utility.