fix(urword,obs): correctly parse boundnames with double quotes
Modify ConstructObservation so that the entire string from the
end of ObsTypeId to the end of the line (including any leading single
or double quotes) is saved in IDstring, which is later parsed by
URWORD into ID1 and ID2. Updated test_gwf_uzf01.py test so that
observation names and boundnames are strings with spaces.