fix(urword,obs): correctly parse boundnames with double quotes
Modify ConstructObservation so that the entire string from the
end of ObsTypeId to the end of the line (including any leading single
or double quotes) is saved in IDstring, which is later parsed by
URWORD into ID1 and ID2. Updated test_gwf_uzf01.py test so that
observation names and boundnames are strings with spaces. Also testing
a quote in a boundname ("left face's").

Modified several autotests to change OC keywords that use ALL, FIRST,
LAST with STEPS or FREQUENCY together. This is not a valid combination,
especially when an integer value is not provided after the STEPS or
FREQUENCY keyword. The missing integer value is now an issue because of
a recent change in FloPy (which results in an endless loop instead of an
ValueError).

black formatting of autotest scripts.