Add conda `environment.yml`
This commit adds an `environment.yml` which should make it easier to keep docs and CI in sync.
It also changes the behavior of the vscode tasks.
It expects an environment called `modflow6` created with the following command.
```
conda env create --force -f environment.yml
```

You might also want to change your `"fortran.fortls.path"` setting to the `fortls` in your `modflow6` environment.

Ideally, you would use this environment for your daily development work as well and adapt it if our dependencies change.