build: refactored meson files for parallel build
Changed meson to allow the MODFLOW 6 exe and lib name to be specified.
By default the names are mf6 and libmf6. This will allow the parallel exe and lib
to have another name (e.g. mf6pro and libmf6pro). Add option to skip build of
mf5to6 and zbud6. Bumped minimum version of meson to 1.1.0. Updated
PARALLEL.md with information on installing OpenMPI and PETSc on MacOS
using Homebrew. Added license for PETSc - this is only added for builds that
use PETSc. Added functionality to use MODFLOW 6 PROFESSIONAL in the
header written to the screen and listing files when MODFLOW 6 is built with
PETSC.