build: refactored meson files for cray build
Refactor meson files to build a CRAY parallel version of MODFLOW using
cray-mpich (7.7.19) and cray-petsc (3.14.5). Bumped minimum version of
meson to 1.1.0. Updated PARALLEL.md with information on installing
OpenMPI and PETSc on MacOS using Homebrew. Added license for PETSc
(this is only added to the license when MODFLOW is built with PETSc.
Added functionality to use MODFLOW 6 PROFESSIONAL in the
header written to the screen and listing files when MODFLOW 6 is built with
PETSC. Added minimal MODFLOW simulation files in .mf6minsim directory
that can be used for testing serial and parallel builds.