ci: remove compiler checks schedule trigger (#1641)
The main purpose of compilers.yml is to check which compiler toolchain/version work on which OS/version. The motivation for the scheduled runs was that github runner images might change underfoot, but this does not impact local development and is only relevant if we want to change how releases are built in CI in future (e.g. adding a gfortran build, using a newer ifort or ifx). So we can just run compiler checks manually whenever needed