refactor(ts): skip search for repeated boundary time series file
The subroutine read_value_or_time_series checks to see if
a time series file (TS) has been previously included in the
time series manager (tsManager) before adding it.

The check compares against the package name and the
input parser row and column (ii, jj) location in the input file.
If the package name and location match, then the TS is not
added to the tsManager.

This check is never true for boundary (BND) type TS,
so a conditional is added to only perform the search for
for non-BND TS types.

This results in a substantial performance increase when there
are a large number of time series files.