ci(parallel): add parallel windows distribution (#1682)
Add a parallel Windows distribution to the release workflow, making it available both on the nightly build and in official releases. The archive is named win64par.zip and contains all non-system dependencies. This avoids the need for software installations or other actions requiring elevation for the end user.

Bundled dependencies include:

- MSVC runtime
  * ucrtbase.dll
  * msvcrt.dll
  * msvcp_win.dll
  * MSVCP140.dll
  * VCRUNTIME140.dll
  * VCRUNTIME140_1.dll
- oneAPI (Fortran compiler, MKL, MPI)
  * libifcoremd.dll
  * libmmd.dll
  * mkl_mc3.2.dll
  * mkl_core.2.dll
  * mkl_avx512.2.dll
  * mkl_avx2.2.dll
  * mkl_def.2.dll
  * mkl_sequential.2.dll
  * libfabric.dll
  * impi.dll
  * mpiexec.exe
  * hydra_bstrap_proxy.exe
  * hydra_pmi_proxy.exe
  * hydra_service.exe
- PETSc
  * libpetsc.dll

Miscellaneous:

- factor out action to build mf6 parallel on windows
- add developmode toggle to release_dispatch.yml