ci: use -Doptimization=1 for arm mac gfortran test and release (#1827)
On rectangular grids, the generalized particle tracking method can find itself in an endless loop on macOS 14 compiled with gfortran with optimization=2, confirmed with gfortran 11-14.

It's difficult to say where things are going wrong since on ARM macs neither gdb nor lldb can be used so we are stuck with print debugging. An architecture/compiler problem seems plausible since we'd presumably see the same behavior on other platforms with opt=2 if it were our bug, but I'm not sure.

This should not occur typically since rectangular cells are detected and solved with Pollock's method, but we use dev_forceternary in test_prt_disv1.py to check that the generalized method reduces to Pollock's. The test has been hanging in CI recently — we might have caught it sooner but there was a separate setup-fortran issue which concealed this one at first.

While it might be safe to release with optimization=2 (with the assumption developers are the only ones using dev options) I don't think we can be sure the same or similar issues could not affect other model grids. So I think we should drop the optimization level pending more information.