fix: build windows parallel with ifort 2021.12 (#1853)
The Windows parallel distribution was built with Intel Fortran Classic 2021.11.1 until now. When running PRT models, Win parallel distributions built with this version of the compiler produce a runtime error in MethodSubcellPool.f90 when the method pool's pointers to tracking methods are deallocated. This seems like a compiler bug, though we also do not properly deallocate all pointers on the method types, which could cause memory leaks if in future PRT is modified such that allocation/deallocation happens multiple times. This will be fixed in a subsequent PR.

The crash only occurs on the Windows parallel distribution because we use ifort 2021.7 to build the other Intel distributions. The reason we did not use 2021.7 to build the Win parallel dist is parallel requires the oneAPI base kit, and I've been unable to find the installer URL for that base kit. Luckily the latest ifort (2021.12.0) doesn't have the same issue, so we can simply switch the Win parallel dist to 2021.12.0.