fix(prt): multiple fixes
The pollock and ternary methods could hang when a particle should terminate due to no exit face. This was because we failed to set the advancing flag to false.

Entangled with this, the ternary method could erroneously terminate a particle and report no exit face (as described above, previously this would hang) due to precision error in the exit time/location calculation when two conditions obtain simultaneously: the particle enters the subcell extremely close to one of its vertices, and flow very nearly parallels one of the subcell's faces. We have encountered similar situations before, solved by nudging the particle some small distance into the interior of the subcell before applying the subcell tracking method. This particular case is resolved by increasing the nudge distance from (machine precision * 100) to (machine precision * 10000). Perhaps in future the nudge distance could be scaled by the size of the subcell.