fix(prt): fix multiple bugs (#1874)
The pollock and ternary methods could hang when a particle should terminate due to no (sub)cell exit face. This occurred because a flag signaling particle advance was not set in the proper location.

Entangled with this, the ternary method could erroneously terminate a particle and report no exit face (as described above, previously this would hang) due to precision error in the exit time/position calculation. This could happen when two conditions are both met: the particle enters the subcell very close to one of its vertices, and flow very nearly parallels one of the subcell's faces. We have encountered similar situations before, solved by nudging the particle a small distance into the interior of the subcell before applying the tracking method. This particular case is resolved by increasing the padding distance from machine precision * 10^2 to machine precision * 10^5.