fix(prt): make extended tracking opt-in
Previously the default behavior was to track particles until termination, as with MODPATH 7 stop time option 'extend'. Under extended tracking, the program may never halt if particles enter a cycle in the flow system. This PR changes the default behavior to the equivalent of MP7 stop time option 'final', terminating particles when the simulation ends unless new PRP keyword option  is provided. This is meant to provide a (stronger) guarantee that the program halts under default settings.

Internally, the extend option is stored as an integer rather than logical in case we need to enumerate more alternatives in future.

This PR also avoids unnecessary iterations in , and changes the default visibility of  members to private, as per style conventions.