fix(prt): sum boundary and face flows when loading cell (#1900)
Previously cell face flows were not properly summed with boundary flows, rather for cell faces with active neighbors, the face flow replaced any boundary flows. The reason this has gone unnoticed til now is probably that IFLOWFACE is typically applied to faces on the boundary of the active domain.

Also factor out a subroutine to load face flows into the cell definition — could be pulled into base MethodType but it is a small amount of duplication and implementations may diverge.