fix(prt): fix pass-to-bottom tracking method (#1918)
The pass-to-bottom method had a reference to an unused cell definition pointer which was moved some time ago to the Cell type. Depending on the platform and compiler, this could result in crashes due to divide by zero errors, or undefined behavior preventing detection of when a particle should exit a cell through its bottom face, leading to early termination.

Autotests for PTB behavior to come in a separate PR. Some testing has been done manually on sample models.