refactor(oc): introduce module/type for time step selection (#1923)
In the spirit of TimeSelect, factor out a module and type TimeStepSelect for time step selections. This simplifies PrintSaveManager which can compose two of the latter. Add unit tests. Cleanup docstrings, comments, and style in base and concrete OC types.

This can also be used by PRT, to come in a separate PR which cleans up particle release logic.