fix(msvs): visual studio project files remove win32 config
The Visual Studio Solution (.sln) and Project (.vfproj) files
no longer include the win32 compilation option. The files
are also updated to Visual Studio 2022 format. This format
still works with an up to date Visual Studio 2019.

The fpp option was set to global for debug and not for release.
It was disabled for debug, and instead added to both
configurations under files that need pre-processing
(those with .F and .F90 extension).

The debug symbols (.pdb) are only generated for debug compilation
and are placed in the obj_temp directory. Debug uses the debug MT,
and release uses the release MT runtime libraries.

The mf6bmi required ignoring the LIBCMT, LIBCMTD, and LIBIFCOREMT
to compile with the 2024 oneAPI. mf6.exe does not require
ignoring these libraries, so LIBCMT is removed as an ignore.

Floating point exceptions are set to fpe0 (underflow is zero,
otherwise raise an exceptions). Heap arrays is enabled for
all configurations and projects.

The intermediate directory was changed from
   $(ProjectName)/$(Platform)/$(ConfigurationName)
to
   $(ProjectName)_$(Platform)_$(ConfigurationName)
to simply the structure from having multiply directories
that contain nothing but a single directory.

Added /Qdiag-disable:10448 to suppress warning about
ifort being deprecated.