fix(msvs): zbud6 and mf5to6 visual studio project are x64
The Visual Studio Solution (.sln) and Project (.vfproj) files
for MF5to6 and Zonebudget are a x64 compilation and the
win32 configuration is removed. The solution and project files
use Visual Studio 2022 format instead of VS2013 format.
This still works with an up to date Visual Studio 2019.

Changed the zonebudget target from zonebudget.exe to zbud6.exe
to reflect what is specified in the meson.build file
(and online downloads).

The mf5to6 and zonebudget projects added a missing
fpp option for successful release compilation. Intermediate
files are now placed in an obj_temp directory and the
resulting executables are placed in the bin folder
(previously, a "copy" post-build command was invoked).

Added /Qdiag-disable:10448 to suppress warning about
ifort being deprecated.