ci: add jobs to build/test with gfortran in debug mode (#1966)
Add jobs to the gfortran test matrix in ci.yml to build and test in debug mode. The intent is to catch situations like the one fixed by #1959. To avoid too much CI burden this PR only adds macos-14 and ubuntu-22.04 jobs. This seems to add ~15min to total runtime.