fix(mf6bmiutil): workaround ARM mac char array assignment issue (#1993)
Assignments from the result of string_to_char_array fail on ARM mac with GNU Fortran when the LHS and RHS are of different lengths. This seems like a compiler bug? Confirmed with gfortran 12-14. Work around it by introducing a subroutine to do the conversion.

@mjr-deltares and I have discussed and determined as a next step to create a minimal reproduction and post on the Fortran discourse forum. I will add a link here once that's done.

This should resolve the modflowapi macOS CI failures.