fix(prp): fix release time selection logic
A conditional expression in the release timing logic was incorrect: we only want to enable the default release time (beginning of the first step of the first period) if no release times are specified via the RELEASETIMES block, the RELEASE_TIME_FREQUENCY option, or period block release time settings.

We also did not previously mention said default. Describe this in the PRP definition file.