fix(prt): convert to local coords in ternary cell method (#2122)
This reverts the switch to calculating the centroid via arithmetic mean for triangular cells in #2119 — the rest of the vertex velocity calculation assumes the polygonal definition of the centroid so that change was an error. A more robust way to avoid intermediate loss of precision is converting to local cell-local coordinates before applying the ternary cell method.