docs(prp): fix removal version of fraction option (#2126)
The PRP DFN described the fraction option has having been removed in 6.5.1, which never materialized. It was removed in 6.6.0 instead.

Fix an mf6io document issue where deprecated/removed options remained in keystring descriptions.

Add cffconvert to the project dependencies, add a corresponding pixi task, and use it in CI. There is some issue with the check cff action.